--- src/main/java/traveler/travel/domain/account/controller/AuthController.java
@@ -1,29 +1,0 @@
-package traveler.travel.domain.account.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import traveler.travel.domain.account.service.AuthService;
-import traveler.travel.global.dto.EmailLoginRequestDto;
-import traveler.travel.global.dto.TokenDto;
-import traveler.travel.global.dto.TokenRequestDto;
-
-@RestController
-@RequestMapping("/auth")
-@RequiredArgsConstructor
-public class AuthController {
-    private final AuthService authService;
-
-    @PostMapping("/login")
-    public ResponseEntity<TokenDto> login(@RequestBody EmailLoginRequestDto dto) {
-        return ResponseEntity.ok(authService.login(dto));
-    }
-
-    @PostMapping("/reissue")
-    public ResponseEntity<TokenDto> reissue(@RequestBody TokenRequestDto tokenRequestDto) {
-        return ResponseEntity.ok(authService.reissue(tokenRequestDto));
-    }
-}
